Fix:: Reservation pacakage refactoring
