+++ src/Project/Reservation/ReservationManager.java
@@ -1,168 +1,198 @@
 package Project.Reservation;
 
-import Project.*;
+import Project.Exception.ExitException;
 import Project.FilesIO.FileDataManager;
+import Project.Movie;
 import Project.Payment.PaymentManager;
+import Project.Reservation.MovieScreeningInfo;
+import Project.Reservation.Reservation;
+import Project.Reservation.ScreeningTime;
+import Project.Schedule;
+import Project.Seat;
 import Project.User.Client;
 import Project.User.User;
 
+import java.time.DateTimeException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.*;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 public class ReservationManager {
+
+    private static final Scanner sc = new Scanner(System.in);
+    private static final String EXIT_COMMAND = "exit";
+    private static String menu;
+    private static String seatInfo;         //좌석 정보 (ex: a1)
+    private static Integer[] seatNumber;    //좌석 정보 (ex: [0, 0])
+
+    private static String selectedDate; //영화 선택 날짜
     // 상영관 및 시간대 정보
     private static String[] theaters = {"1관", "2관", "3관"};
     private static ScreeningTime[] times = {ScreeningTime.time_09, ScreeningTime.time_12, ScreeningTime.time_18};
 
-    //1) 예매하기
-    public static void makeMovieReservation(User user) {
-        Scanner sc = new Scanner(System.in);
+    public  static void start(User user) {
+        //예매 하기 출력
+        printReservationStat();
 
-        Map<String, Map<Movie,Schedule[][]>> data = FileDataManager.readMovieScheduleFromFile();    // MovieSchedule.txt
+        //파일 데이터 불러오기
+        Map<String, Map<Movie, Schedule[][]>> data = FileDataManager.readMovieScheduleFromFile();    // MovieSchedule.txt
         List<User> usersList = FileDataManager.readUserInfoFromFile();                              //유저정보읽기
         List<Reservation> reservationsList = FileDataManager.readReservationsFromFile();            //예매 정보 읽기
 
-        //유저 정보 최신화
+        //유저 정보 최신화 (예왜처리 불필요)
+        Client client = (Client) getUserInfo(user, usersList);
+
+        //날짜 문자열만 관리
+        Set<String> scheduleDatesSet = data.keySet();
+
+        try {
+            //데이터에서 현재 날짜 이후의 예매 가능 날짜 출력
+            displayScheduleDates(scheduleDatesSet);
+
+            //1. 사용자에게 영화 상영 날짜를 입력받는다.
+            inputSelectDate(scheduleDatesSet);
+
+            //2. 입력받은 날짜의 상영 정보를 저장하고 보여준다.
+            Map<Integer, MovieScreeningInfo> scheduleNumbersMap = getAndPrintMovieSchedules(data);      //영화 상영 스케쥴에 넘버링한 자료구조(메인 데이터 관리)
+
+            //3.사용자에게 원하는 영화 스케쥴을 입력받는다.
+            Integer choiceSchedule = inputChoiceSchedule(scheduleNumbersMap);
+
+            // 필요한 데이터 저장 (선택한 영화, 상영관 번호, 상영 시간 번호, 상영관 좌석 정보)
+            Movie choiceMovie = scheduleNumbersMap.get(choiceSchedule).getMovie();  //선택한 영화 정보
+            int row = scheduleNumbersMap.get(choiceSchedule).getI();              //선택한 상영관
+            int col = scheduleNumbersMap.get(choiceSchedule).getJ();              //선택한 시간
+            Seat[][] seats = scheduleNumbersMap.get(choiceSchedule).getSchedule().getSeats();   //이걸로 변경
+
+            //4 좌석 선택
+            printOneLine(new String[]{choiceMovie.getTitle(), String.valueOf(times[col]), theaters[row]});
+            choiceSeatProcess(seats);
+
+            //5. 결제 여부 확인
+            checkPaymentMovie();
+
+            //6. 결제 (결제 실패시 메뉴로 이동)
+            moviePaymentProcess(choiceMovie, client);
+
+            //7. 예매 정보 저장
+            updateScheduleInfo(scheduleNumbersMap, choiceSchedule);
+
+            //8. 예매 내역 생성
+            //8.1 예매 번호 생성
+            String ticketNumber = ticketNumberGenerator(row, col);
+
+            //8.2 예매 정보 생성
+            Reservation myReservation = Reservation.builder()
+                    .reservationDate(LocalDateTime.now())
+                    .reservationNumber(ticketNumber)
+                    .user(client)
+                    .movie(choiceMovie)
+                    .movieDate(LocalDate.parse(selectedDate, DateTimeFormatter.ofPattern("yyyy-MM-dd")))
+                    .theater(new Integer[] {row, col})
+                    .seat(seats[seatNumber[0]][seatNumber[1]])
+                    .moviePrice(choiceMovie.getPrice())
+                    .build();
+
+            //8 파일 덮어쓰기
+            FileDataManager.writeMovieScheduleToFile(data);     //MovieSchedule.txt
+            writeTextData(reservationsList, myReservation);     //Reservations.txt
+            writeTextData(usersList, myReservation, client);    //Users.txt
+
+
+        } catch (ExitException e) {
+            System.out.println(e.getMessage());
+        } finally {
+            menu = null;
+            seatInfo = null;
+            seatNumber = null;
+            selectedDate = null;
+        }
+    }
+
+    ///////////////////////////////////////   Service   /////////////////////////////////////////////////////
+    //유저 정보 가져오기
+    private static User getUserInfo(User user, List<User> usersList) {
         Optional<User> foundUserOptional  = usersList.stream()
                 .filter(findUser -> findUser.getEmail().equals(user.getEmail()))
                 .findFirst();
-        Client client = (Client) foundUserOptional.orElse(null);
 
-        //영화 스케쥴 넘버링(메인 데이터 관리)
-        Map<Integer, MovieScreaningInfo> scheduleNumbersMap = new HashMap<Integer, MovieScreaningInfo>();
+        return foundUserOptional.orElse(null);
+    }
+
+    private static String checkInputMenu() {
+        menu = sc.nextLine();
+        if(menu.equals(EXIT_COMMAND)) throw new ExitException();
+        return menu;
+    }
 
 
-        //1. 사용자에게 영화 상영 날짜를 입력받는다. (while 문)
+    private static void checkInputSeatInfo() {
+        seatInfo = sc.nextLine();
+        if(seatInfo.equals(EXIT_COMMAND)) throw new ExitException();
+    }
 
-        String selectedDate = ReservationException.DataInputFormat();
-        if (selectedDate == null) {
-            return; // 사용자가 다시 입력하지 않기로 결정한 경우 메서드 종료
+    private static void displayScheduleDates(Set<String> scheduleDatesSet) {
+        List<LocalDate> sortedDates = new ArrayList<>();
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+
+        for (String dateString : scheduleDatesSet) {
+            LocalDate localDate = LocalDate.parse(dateString, formatter);
+            sortedDates.add(localDate);
         }
+        Collections.sort(sortedDates);
+        printSortedMovieScheduleDates(sortedDates);
+    }
 
-        ReservationException.checkScheduleExistence(selectedDate, data);
-        if (data.get(selectedDate) == null) {
-            // 스케줄이 존재하지 않으면 다시 날짜를 입력받음
-            makeMovieReservation(user);
-            return; // 재귀 호출 후 메서드 종료
+
+    //예매를 선택하는 날짜 입력받기
+    private static void inputSelectDate(Set<String> scheduleDatesSet) {
+        do {
+            printInputMessage("영화 상영 날짜를 입력하세요 [ex) 2024-03-23] : ");
+            selectedDate = sc.nextLine();
+            if (!isValidationDate(selectedDate) || !isLocalDateFormat(selectedDate)) {
+                //잘못 입력하셨습니다.
+                printInvalidInputMessage("유효하지 않은 날짜형식입니다.");
+                checkInputMenu();
+
+            } else if (!scheduleDatesSet.contains(selectedDate)) {
+                //상영 가능한 날짜 없습니다.
+                printInvalidInputMessage("해당 날짜에는 상영 정보가 없습니다.");
+                checkInputMenu();
+
+            } else {
+                break;
+            }
+        } while(true);
+    }
+
+    //날짜 문자열 유효성 검사
+    private static boolean isValidationDate(String inputDate) {
+        String regex = "\\d{4}-\\d{2}-\\d{2}";
+        Pattern pattern = Pattern.compile(regex);
+        Matcher matcher = pattern.matcher(inputDate);
+        return matcher.matches();
+    }
+
+    //날짜 문자열 LocalDate 타입 변환 검증
+    private static boolean isLocalDateFormat(String inputDate) {
+        try {
+            LocalDate localDate = LocalDate.parse(inputDate, DateTimeFormatter.ofPattern("yyyy-MM-dd"));
+            return true;
+        } catch (DateTimeException e) {
+            return false;
         }
-        //1.1 입력 날짜 검증 필요 (형식이 올바른지)
-        //1.2 입력 날짜 스케쥴 존재 여부 검증 필요
-
-        Map<Movie, Schedule[][]> movieMap = data.get(selectedDate); //선택한 날짜의 영화상영정보
-        System.out.println("선택하신 날짜 : " + selectedDate);
-        System.out.println("------------------------------------");
-
-        //2. 입력받은 날짜의 상영 정보를 보여준다.
-        System.out.println("=====영화 상영 일정====");
-        getMovieSchedules(movieMap, scheduleNumbersMap);
-
-
-        //3.사용자에게 원하는 영화 스케쥴을 입력받는다. (while 문)
-        //
-        // 올바른 입력 확인
-        Integer choiceSchedule = ReservationException.choiceSchedule(  scheduleNumbersMap);
-
-
-        if (choiceSchedule ==null) {
-            System.out.println("메뉴로 돌아갑니다");
-            return;
-        }
-        // 스케쥴 존재 확인
-        //3-1
-       // boolean check = ReservationException.checkInputScheduleExistence(choiceSchedule,scheduleNumbersMap);
-        //상영관 검증 로직 필요 (존재 여부, 올바른 입력 확인)
-        /*
-        if(check == false) {
-            System.out.println("메뉴로 돌아갑니다.");
-            return;
-        }
-
-         */
-
-        Movie choiceMovie = scheduleNumbersMap.get(choiceSchedule).getMovie();  //선택한 영화 정보
-        Integer row = scheduleNumbersMap.get(choiceSchedule).getI();              //선택한 상영관
-        Integer col = scheduleNumbersMap.get(choiceSchedule).getJ();              //선택한 시간
-        System.out.println("영화 : " + choiceMovie.getTitle());
-        System.out.printf("%s, (%s)\n", ReservationManager.times[col], ReservationManager.theaters[row]);
-
-        //4 좌석 선택 (while 문)
-        //4.1 해당 상영관의 좌석을 보여주기
-        Seat[][] seats = scheduleNumbersMap.get(choiceSchedule).getSchedule().getSeats();   //이걸로 변경
-//        Seat[][] seats = movieMap.get(choiceMovie)[row][col].getSeats();
-        getSeats(seats);
-
-        //4.2 사용자에게 좌석을 입력받기
-        //4.2 좌석 번호 선택 이력 검증 로직(a1 또는 A1 형식으로 받도록)
-        String selectSeat = ReservationException.validateSeatSelection();  //선택한 좌석번호
-        if (selectSeat == null) {
-            System.out.println("메뉴로 돌아갑니다");
-            return;
-        }
-        Integer[] seatNumber = convertSelectSeat(selectSeat);   // 선택한 좌석번호 변환 -> 좌석 번호 [행, 열]
-        //4.3 예매 좌석 검증 필요 (null인지 아닌지 여부(예매 여부), 올바른 입력 여부)
-
-        boolean seatCheck = ReservationException.checkEmptySeat(seatNumber,scheduleNumbersMap, choiceSchedule);
-        if(seatCheck == false) {
-            System.out.println("메뉴로 돌아갑니다.");
-            return;
-        }
-
-        //5. 결제
-        //결제 여부 확인 필요
-        Integer moviePrice = choiceMovie.getPrice();
-        if (!PaymentManager.payPoint(client, moviePrice)) {     //결제 실패시
-            System.out.println("메뉴화면으로 돌아갑니다.");
-            return;
-        }
-
-        //6. 예매 정보 저장 (결제 성공시)
-        //6.1 좌석 정보를 저장한다.
-        seats[seatNumber[0]][seatNumber[1]] = new Seat(seatNumber[0], seatNumber[1]);   //예매한 좌석 정보 저장
-        scheduleNumbersMap.get(choiceSchedule).getSchedule().seatCountDown();           //빈좌석 - 1
-
-        //6.2 예매번호 생성 후 저장
-        String ticketNumber = ticketNumberGenerator(selectedDate, row, col ,selectSeat);
-//        System.out.println("예매가 완료되었습니다.");
-//        System.out.printf("예매번호 : [%s]\n", ticketNumber);
-
-        //6.3 예매정보 내역 저장 -> Reservation
-        LocalDate choiceMovieDate = LocalDate.parse(selectedDate, DateTimeFormatter.ofPattern("yyyy-MM-dd"));
-
-        Reservation myReservation = Reservation.builder()
-                .reservationDate(LocalDateTime.now())
-                .reservationNumber(ticketNumber)
-                .user(client)
-                .movie(choiceMovie)
-                .movieDate(choiceMovieDate)
-                .theater(new Integer[] {row, col})
-                .seat(seats[seatNumber[0]][seatNumber[1]])
-                .moviePrice(choiceMovie.getPrice())
-                .build();
-
-        OutputView.movietiketPrint(myReservation);
-        //7. 파일 덮어쓰기
-        //7.1 MovieSchedule.txt
-        FileDataManager.writeMovieScheduleToFile(data);
-
-        //7.2 Reservations.txt
-        if (reservationsList == null) {
-            reservationsList = new LinkedList<>();
-        }
-        reservationsList.add(myReservation);
-        FileDataManager.writeReservationToFile(reservationsList);
-
-        //7.3 Users.txt
-        client.getReservationList().add(myReservation);
-//        usersList.stream().filter(finduser -> finduser.getEmail().equals(client.getEmail()))
-//                        .findFirst()
-//                        .ifPresent(finduser -> ((Client) finduser).getReservationList().add(myReservation));
-        FileDataManager.writeUserInfoToFile(usersList);
     }
 
     //입력받은 날짜의 상영정보를 보여준다
-    public static void getMovieSchedules(Map<Movie, Schedule[][]> movieMap, Map<Integer, MovieScreaningInfo> scheduleNumbersMap ) {
+    private static Map<Integer, MovieScreeningInfo> getAndPrintMovieSchedules(Map<String, Map<Movie, Schedule[][]>> data) {
+        Map<Integer, MovieScreeningInfo> scheduleNumbersMap = new HashMap<>(); //날짜 변경에 따른 초기화
+        Map<Movie, Schedule[][]> movieMap = data.get(selectedDate);     //선택한 날짜의 영화상영정보
+
+        System.out.printf("======== [ %s ] 영화 상영 일정] ========\n", selectedDate);
+
         //영화 순서 정렬
         List<Map.Entry<Movie, Schedule[][]>> movieEntryList = new ArrayList<>(movieMap.entrySet());
         movieEntryList.sort(Comparator.comparing(entry -> entry.getKey().getTitle()));
@@ -171,7 +201,7 @@ public class ReservationManager {
         for (Map.Entry<Movie, Schedule[][]> entry : movieEntryList) {
             System.out.println("제목 : " + entry.getKey().getTitle());
             System.out.print("       ");
-            for (ScreeningTime time : ReservationManager.times) {
+            for (ScreeningTime time : times) {
                 System.out.printf("%10s", time);
             }
             System.out.println();
@@ -186,70 +216,132 @@ public class ReservationManager {
                     int emptySeats = entry.getValue()[i][j].getEmpty();     //빈좌석
                     int totalSeats = entry.getValue()[i][j].getTotal();     //총좌석
                     String seatInfo = emptySeats + "/" + totalSeats;
-                    scheduleNumbersMap.put(scheduleNumber, new MovieScreaningInfo(entry.getKey(), i, j, entry.getValue()[i][j]));   //schedule 넘버
+                    scheduleNumbersMap.put(scheduleNumber, new MovieScreeningInfo(entry.getKey(), i, j, entry.getValue()[i][j]));   //schedule 넘버
                     System.out.printf("[%d]\t%s |", scheduleNumber++, seatInfo);
                 }
                 System.out.println();
                 System.out.print("       ");
                 for (int j = 0; j < times.length; j++) {
-                    System.out.print("------------");
+                    System.out.print("----------");
                 }
                 System.out.println();
             }
             System.out.println();
         }
+        return scheduleNumbersMap;
     }
 
-    //좌석 조회하기
-    public static void getSeats(Seat[][] seats) {
-        // 좌석 맵 출력
-        System.out.println("============= 좌석 정보 ============");
-        System.out.println("============= SCREEN =============");
-        System.out.print("   ");
-
-        // 열 번호 출력
-        for (int j = 0; j < seats[0].length; j++) {
-            System.out.printf("%-6s", "  " + (j + 1));
-        }
-        System.out.println();
-
-
-        for (int i = 0; i < seats.length; i++) {
-            System.out.print(((char) (i + 65)) + " "); // 행 알파벳 출력
-            for (int j = 0; j < seats[0].length; j++) {
-                if (seats[i][j] == null) {
-                    System.out.printf("%-6s", "[ " + ((char) (i + 65)) + "" + (j + 1) + " ]");
-                } else {
-                    System.out.printf("%-6s", "[ X ]"); // 예약된 좌석
-                }
+    //3.사용자에게 원하는 영화 스케쥴을 입력받는다.
+    private static Integer inputChoiceSchedule(Map<Integer, MovieScreeningInfo> scheduleNumbersMap) {
+        do {
+            printInputMessage("원하는 영화의 스케쥴을 선택해 주세요. [ex) 38]    > ");
+            String inputData = sc.nextLine();
+            if (!inputData.matches("\\d+")) {
+                //숫자형식으로 입력하지 않은 경우
+                printInvalidInputMessage("숫자만 입력받습니다.");
+                checkInputMenu();
+            } else if (!scheduleNumbersMap.containsKey(Integer.parseInt(inputData))) {
+                //상영 스케쥴에 존재하지 않는 번호인 경우
+                printInvalidInputMessage("해당 번호에는 배치된 영화가 없습니다.");
+                checkInputMenu();
+            } else {
+                return Integer.parseInt(inputData);
             }
-            System.out.println();
-        }
-        System.out.println("=================================");
-        System.out.println("=================================");
+        } while(true);
     }
 
-    //사용자가 입력한 좌석 번호(알파벳 + 숫자 조합)를 행, 열 정수형으로 변환
-    public static Integer[] convertSelectSeat (String selectSeat) {
-        selectSeat = selectSeat.trim().toUpperCase();
+    //4. 좌석 선택
+    private static void choiceSeatProcess(Seat[][] seats) {
+        outLoop:
+        do {
+            //해당 상영관의 좌석 출력
+            printMovieScheduleSeats(seats);
+
+            //좌석을 입력받기
+            do {
+                printInputMessage("좌석을 선택하세요. [ex) A1 또는 a1]    > ");
+                checkInputSeatInfo();
+                if (!isValidationSeat()) {
+                    //유효한 형식이 아닌 경우
+                    printInvalidInputMessage("올바른 입력값이 아닙니다.");
+                    checkInputSeatInfo();
+
+                } else {
+                    convertSelectSeat();    //seatinfo -> seatNumber 로 변환
+                    if (!isExistSeat(seats)) {
+                        printInvalidInputMessage("해당 좌석은 존재하지 않습니다.");
+                        checkInputSeatInfo();
+
+                    } else if (!isCheckEmptySeat(seats)) {
+                        //이미 예매된 경우
+                        printInvalidInputMessage("해당 좌석은 이미 예매되었습니다");
+                        checkInputSeatInfo();
+                    } else {
+                        break outLoop;
+                    }
+                }
+            } while (true);
+        } while (true);
+    }
+
+    //좌석 번호 선택 이력 검증 로직(a1 또는 A1 형식으로 받도록)
+    private static boolean isValidationSeat() {
+        String seatPattern = "[A-Za-z][0-9]+"; // 알파벳 + 숫자 형식으로 구성되어야 함
+        return seatInfo.matches(seatPattern);
+    }
+
+    //좌석 번호가 선택 범위를 벗어나는 경우
+    private static boolean isExistSeat(Seat[][] seats) {
+        return ((0 <= seatNumber[0]) && (seats.length > seatNumber[0]) && (0 <= seatNumber[1]) && (seats[0].length > seatNumber[1]));
+    }
+
+
+    //좌석 번호가 이미 예매된 경우
+    public static boolean isCheckEmptySeat(Seat[][] seats) {
+        return (seats[seatNumber[0]][seatNumber[1]] == null);
+    }
+
+    //좌석 변환 (seatinfo -> seatNumber 로 변환)
+    private static void convertSelectSeat () {
+        seatInfo = seatInfo.trim().toUpperCase();
 
         // 알파벳과 숫자 분리
         String rowStr = ""; // 알파벳
         String colStr = ""; // 숫자
-        for (char c : selectSeat.toCharArray()) {
-            if (Character.isLetter(c)) {
-                rowStr += c;
-            } else if (Character.isDigit(c)) {
-                colStr += c;
-            }
+        for (char c : seatInfo.toCharArray()) {
+            if (Character.isLetter(c)) rowStr += c;
+            else if (Character.isDigit(c)) colStr += c;
         }
-        Integer row = rowStr.charAt(0) - 'A';
-        Integer col = Integer.parseInt(colStr) - 1;
-        return new Integer[] {row, col};
+        seatNumber = new Integer[] {rowStr.charAt(0) - 'A', Integer.parseInt(colStr) - 1};
     }
 
-    //예매번호 생성기 (16자리)
-    public static String ticketNumberGenerator(String selectedDate, Integer row, Integer col, String selectSeat) {
+    private static void moviePaymentProcess(Movie choiceMovie,Client client) {
+        Integer moviePrice = choiceMovie.getPrice();
+        if (!PaymentManager.payPoint(client, moviePrice)) throw new ExitException();
+    }
+
+
+    private static void checkPaymentMovie() {
+        printInputMessage("해당 예매를 확정하시겠습니까? [1] 예매 확정 / [exit] 예매 취소    > ");
+        do {
+            menu = sc.nextLine();
+            if(menu.equals(EXIT_COMMAND)) throw new ExitException();
+            if(!menu.equals("1")) {
+                printInvalidInputMessage("잘못된 입력입니다. 다시 입력해주시길 바랍니다.");
+            } else {
+                return;
+            }
+        } while(true);
+    }
+
+    private static void updateScheduleInfo(Map<Integer, MovieScreeningInfo> scheduleNumbersMap, Integer choiceSchedule) {
+        //좌석 정보 업데이트
+        scheduleNumbersMap.get(choiceSchedule).getSchedule().getSeats()[seatNumber[0]][seatNumber[1]] = new Seat(seatNumber[0], seatNumber[1]);
+        scheduleNumbersMap.get(choiceSchedule).getSchedule().seatCountDown();           //빈좌석 - 1
+
+    }
+
+    public static String ticketNumberGenerator(Integer row, Integer col) {
         StringBuilder sb = new StringBuilder();
         Random random = new Random();
         //1. 날짜(2024-03-23) -> 10진수 (20240323) -> 16진수로 변환
@@ -263,7 +355,7 @@ public class ReservationManager {
         sb.append(hexString).append("-");
 
         //2. 상영관, 상영시각, 좌석번호
-        sb.append(row).append(col).append(selectSeat).append("-");
+        sb.append(row).append(col).append(seatInfo).append("-");
 
         //3. 8자리는 알파벳, 숫자의 랜덤 조합
         for (int i = 0; i < 8; i++) {
@@ -291,253 +383,83 @@ public class ReservationManager {
         }
         return sb.toString().toUpperCase();
     }
-    ////////////////////////////////////////////////////////////////////////////////////////////////
-    //2) 예매 조회하기
-    public static void getReservation(User user) {
-        Scanner sc = new Scanner(System.in);
-        List<User> usersList = FileDataManager.readUserInfoFromFile(); // User.txt
 
-        Optional<User> foundUserOptional  = usersList.stream()
-                                            .filter(findUser -> findUser.getEmail().equals(user.getEmail()))
-                                            .findFirst();
-
-        Client client = (Client) foundUserOptional.orElse(null);
-        List<Reservation> userReservationsList = client.getReservationList();
-
-        String menu = "";
-        do {
-            System.out.println("전체 예매조회 - [1]");
-            System.out.println("기간별 예매조회 - [2]"); //영화를 예매한 날짜에 대해
-            System.out.println("영화별 전체조회 - [3]");
-            System.out.println("메뉴로 돌아가기 - [0]");
-            System.out.print("메뉴를 선택해주세요 : ");
-            menu = sc.nextLine();
-
-            switch (menu) {
-                case "1":   //전체 예매 조회
-                    getAllReservations(userReservationsList);
-                    break;
-                case "2":
-                    //기간별 예매조회
-                    String periodMenu = "";
-                    do {
-                        System.out.println("최근 1개월 조회 - [1]");
-                        System.out.println("지정 기간 조회 - [2]");
-                        System.out.println("뒤로가기 - [-1]");
-                        System.out.println("메뉴로 돌아가기 - [0]");
-                        periodMenu = sc.nextLine();
-                        //예외처리
-
-                        String[] findPeriod = new String[2];    //조회 기간 [시작 날짜, 끝날짜]
-                        String findDate = "";
-                        //한달 전 기간 찾기
-                        switch (periodMenu) {
-                            case "1":
-                                DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-                                findPeriod[0] = LocalDate.now().minusMonths(1).format(formatter);   //현재 날짜 기준 한달전
-                                findPeriod[1] = LocalDate.now().format(formatter);  //현재 날짜
-                                getReservationsByPeriod(userReservationsList, findPeriod);   //해당 기간의 예매 정보 출력
-                                break;
-                            case "2":
-                                System.out.println("시작 날짜 입력 ex) 2024-03-23");
-                                findDate = sc.nextLine();
-                                //날짜 예외 처리 필요 -> 예외처리 후 저장
-                                findPeriod[0] = findDate;
-
-                                System.out.println("끝 날짜 입력 ex) 2024-03-23");
-                                findDate = sc.nextLine();
-                                //날짜 예외 처리 필요 -> 예외처리 후 저장
-                                findPeriod[1] = findDate;
-                                getReservationsByPeriod(userReservationsList, findPeriod);   //해당 기간의 예매 정보 출력
-                                break;
-                            case "-1":
-                                System.out.println("예매 조회 메뉴로 돌아갑니다.");
-                                break;
-                            case "0":
-                                System.out.println("메인 메뉴로 돌아갑니다.");
-                                menu = periodMenu;
-                                break;
-                            default:
-                                System.out.println("잘못된 입력입니다.");
-                        }
-                    } while(!(periodMenu.equals("0") || periodMenu.equals("-1")));
-                    break;
-                //getReservationsByPeriod(reservationList);
-                //기간별 조회는 최근 1개월, 사용자 기간 지정
-                case "3":
-                    //영화의 예매조회
-                    String movieTitleToFilter = "";
-                    System.out.printf("영화 제목 입력 ex) 서울의봄 : ");
-                    movieTitleToFilter = sc.nextLine();
-                    getReservationsForMovie(userReservationsList, movieTitleToFilter);
-                    break;
-                case "0":
-                    System.out.println("메뉴로 돌아갑니다.\n");
-                    return;
-                default:
-                    System.out.println("올바른 입력이 아닙니다.\n");
-            }
-        } while (!menu.equals("0"));
+    private static void writeTextData(List<Reservation> reservationsList, Reservation myReservation) {
+        if (reservationsList == null) {
+            reservationsList = new LinkedList<>();
+        }
+        reservationsList.add(myReservation);
+        FileDataManager.writeReservationToFile(reservationsList);
     }
 
-    //전체 예매 조회
-    private static void getAllReservations(List<Reservation> reservationList) {
-        System.out.println("==================전체 예매 조회====================");
-        for (Reservation reservation : reservationList) {
-            System.out.println(reservation);
+    private static void writeTextData(List<User> usersList, Reservation myReservation, Client client) {
+        client.getReservationList().add(myReservation);
+        FileDataManager.writeUserInfoToFile(usersList);
+    }
+
+
+    //////////////////////////////   OutputPrint   //////////////////////////////////////
+    public static void printInputMessage(String message) {
+        System.out.print(message);
+    }
+
+    public static void printInvalidInputMessage(String message) {
+        System.out.println();
+        System.out.printf("%s [1] 다시 입력 / [exit] 메뉴로 돌아가기\n", message);
+    }
+
+    public static void printOneLine(String[] strArr) {
+        for (String str : strArr) {
+            System.out.printf("%s\t", str);
         }
-        System.out.println("=================================================");
         System.out.println();
     }
 
-    //기간 별 예매 조회(예매 날짜 기준)
-    private static void getReservationsByPeriod(List<Reservation> reservationList, String[] findPeriod) {
+    public static void printReservationStat() {
+        System.out.println("===============================================");
+        System.out.println("==================  예매 하기  ==================");
+        System.out.println("===============================================");
+        System.out.println();
+    }
+
+    public static void printSortedMovieScheduleDates(List<LocalDate> sortedDates) {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-        LocalDate startDate = LocalDate.parse(findPeriod[0], formatter);
-        LocalDate endDate = LocalDate.parse(findPeriod[1], formatter);
-
-        List<Reservation> filteredReservationList = reservationList.stream()
-                .filter(reservation -> {
-                    LocalDate reservationDate = reservation.getReservationDate().toLocalDate();
-                    return (reservationDate.isEqual(startDate) || reservationDate.equals(endDate) ||
-                            (reservationDate.isAfter(startDate) && reservationDate.isBefore(endDate)));
-                })
-                .toList();
-
-        // 필터링된 예약 리스트 출력
-        System.out.printf("=== 예약 목록 (%s - %s) ===\n", findPeriod[0], findPeriod[1]);
-        if (filteredReservationList.isEmpty()) {
-            System.out.println("해당 가긴의 예매 내역을 찾을 수 없습니다.");
-        } else {
-            for (Reservation reservation : filteredReservationList) {
-                System.out.println(reservation);
-            }
+        System.out.print("영화 상영 날짜 : ");
+        for (LocalDate localDate : sortedDates) {
+            String formattedDate = localDate.format(formatter);
+            System.out.printf("\t[%s]", formattedDate);
         }
-        System.out.println("=================================================");
         System.out.println();
     }
 
-    //특정 영화 예매 조회
-    private static void getReservationsForMovie(List<Reservation> reservationList, String movieTitleToFilter) {
-        List<Reservation> filteredReservationList = reservationList.stream()
-                .filter(reservation -> reservation.getMovie().getTitle().equals(movieTitleToFilter))
-                .toList();
-        // 필터링된 예약 리스트 출력
-        System.out.println("=== " + movieTitleToFilter + " 예약 목록 ===");
-        if (filteredReservationList.isEmpty()) {
-            System.out.println("해당 영화의 예매 내역을 찾을 수 없습니다.");
-        } else {
-            for (Reservation reservation : filteredReservationList) {
-                System.out.println(reservation);
-            }
+    public static void printMovieScheduleSeats(Seat[][] seats) {
+        // 좌석 맵 출력
+        System.out.println("============= 좌석 정보 ============");
+        System.out.println("============= SCREEN =============");
+        System.out.print("   ");
+
+        // 열 번호 출력
+        for (int j = 0; j < seats[0].length; j++) {
+            System.out.printf("%-6s", "  " + (j + 1));
         }
-        System.out.println("=================================================");
         System.out.println();
-    }
 
 
-    ////////////////////////////////////////////////////////////////////////////////////////////////
-    //3) 예매 취소하기
-    public static void deleteReservation(User user) {
-        Scanner sc = new Scanner(System.in);
-        Map<String, Map<Movie,Schedule[][]>> data = FileDataManager.readMovieScheduleFromFile();    // MovieSchedule.txt
-        List<User> usersList = FileDataManager.readUserInfoFromFile();                              // User.txt
-        List<Reservation> reservationsList = FileDataManager.readReservationsFromFile();            // reservationsList
-
-        //현재 유저 정보 가져오기
-        Optional<User> foundUserOptional  = usersList.stream()
-                .filter(findUser -> findUser.getEmail().equals(user.getEmail()))
-                .findFirst();
-
-        Client client = (Client) foundUserOptional.orElse(null);
-
-        //현재 유저의 예매 목록
-        List<Reservation> userReservationsList = client.getReservationList();
-        getAllReservations(userReservationsList);   //목록 보여주기
-
-        String check = "";
-        //유저에게 삭제할 예매번호 입력받기
-        do {
-            System.out.print("예매 취소를 위해 예매번호를 입력해주세요: ");
-            String inputReservationNumber = sc.nextLine();
-
-            Optional<Reservation> foundReservationOptional = userReservationsList.stream()
-                    .filter(reservation -> reservation.getReservationNumber().equals(inputReservationNumber))
-                    .findFirst();
-
-            Reservation foundReservation = foundReservationOptional.orElse(null);
-            System.out.println(foundReservation.getMoviePrice());
-
-            //예외 처리
-            if (foundReservation == null) {
-                System.out.println("잘못된 예매 번호 입니다. 다시입력하시겠습니까? (Y / N)");
-                check = sc.nextLine();
-                if (check.equals("Y")) continue;
-                else return;
-            }
-
-            //예외 처리 - 현재 날짜보다 지난 영화의 경우
-//            if (foundReservation.getMovieDate().isBefore(LocalDate.now())) {
-//                System.out.println("기간이 지난 예매내역으로 취소하실 수 없습니다.");
-//                System.out.println("다시입력하시겠습니까? (Y / N)");
-//                check = sc.nextLine();
-//                if (check.equals("Y")) continue;
-//                else return;
-//            } else if (foundReservation.getMovieDate().isEqual(LocalDate.now())) {
-//                LocalDateTime[] screeningTime = {
-//                        LocalDateTime.of(foundReservation.getMovieDate(), LocalTime.of(9, 0)),
-//                        LocalDateTime.of(foundReservation.getMovieDate(), LocalTime.of(12, 0)),
-//                        LocalDateTime.of(foundReservation.getMovieDate(), LocalTime.of(18, 0))
-//                };
-//
-//                if (screeningTime[foundReservation.getTheater()[1]].isBefore(LocalDateTime.now())) {
-//                    System.out.println("기간이 지난 예매내역으로 취소하실 수 없습니다.");
-//                    System.out.println("다시입력하시겠습니까? (Y / N)");
-//                    check = sc.nextLine();
-//                    if (check.equals("Y")) continue;
-//                    else return;
-//                }
-//            }
-            ////위 부분은 사용 예정
-
-            System.out.println(foundReservation);
-            System.out.print("정말로 예매를 취소하시겠습니까? (Y / N)");
-            check = sc.nextLine();
-            if (check.equals("N")) return;
-
-            //MovieSchedule.txt
-            String selectedDate = foundReservation.getMovieDate().format(DateTimeFormatter.ofPattern("yyyy-MM-dd"));
-            Movie choiceMovie = foundReservation.getMovie();        //선택한 영화
-            Integer I = foundReservation.getTheater()[0];           //상영관 번호
-            Integer J = foundReservation.getTheater()[1];           //상영 시각
-            Integer seatI = foundReservation.getSeat().getRow();    //좌석 행
-            Integer seatJ = foundReservation.getSeat().getCol();    //좌석 열
-
-            for (Map.Entry<Movie, Schedule[][]> entry : data.get(selectedDate).entrySet()) {
-                if (entry.getKey().getTitle().equals(choiceMovie.getTitle())) {
-                    Schedule choiceSchedule = entry.getValue()[I][J];
-                    choiceSchedule.seatCountUp();                       //빈좌석 + 1
-                    choiceSchedule.getSeats()[seatI][seatJ] = null;     //좌석 null로 변경
-                    break;
+        for (int i = 0; i < seats.length; i++) {
+            System.out.print(((char) (i + 65)) + " "); // 행 알파벳 출력
+            for (int j = 0; j < seats[0].length; j++) {
+                if (seats[i][j] == null) {
+                    System.out.printf("%-6s", "[ " + ((char) (i + 65)) + "" + (j + 1) + " ]");
+                } else {
+                    System.out.printf("%-6s", "[ X ]"); // 예약된 좌석
                 }
             }
-            FileDataManager.writeMovieScheduleToFile(data);
-
-            //Reservations.txt
-            reservationsList.remove(foundReservation);
-            FileDataManager.writeReservationToFile(reservationsList);
-
-            //User.txt
-            //지불한 포인트 환불
-            Integer refundMoviePrice = foundReservation.getMoviePrice();
-            PaymentManager.payBack(client, refundMoviePrice);
-            System.out.printf("[%d] 포인트가 반환되었습니다.\n", refundMoviePrice);
-            System.out.printf("현재 포인트는 [%d] 입니다\n", client.getPoint());
-
-            client.getReservationList().remove(foundReservation);
-            FileDataManager.writeUserInfoToFile(usersList);
-            break;
-
-        } while(true);
+            System.out.println();
+        }
+        System.out.println("=================================");
+        System.out.println("=================================");
     }
+
+
+
 }
